Inloggen werkt
Inloggen werkt volledig.
Om er zijn twee gebruikers aangemaakt:
john.doe@example.com
PassJohn
(Dit is een manager en een systeem beheerder)

jane.smith@example.com
PassJane
(Deze is geen manager en geen systeem beheerder)
