+++ bumbo/Views/Account/Login.cshtml
@@ -10,7 +10,7 @@
 <!-- Fullscreen container to center content -->
 <div class="min-h-screen flex items-center justify-center bg-gray-100">
     <div class="w-full max-w-sm bg-white shadow-md rounded-lg p-8">
-        
+
         <!-- Icon (optional, if you have an image or icon) -->
         <div class="flex justify-center mb-4">
             <img src="/media/bumbo-logo.png" alt="icon" class="w-24 h-auto">
@@ -19,13 +19,24 @@
         <!-- Title -->
         <h2 class="text-2xl font-semibold text-center text-gray-800 mb-6">Log in je account</h2>
 
+        <!-- Error message section -->
+        @if (!ViewData.ModelState.IsValid)
+        {
+            <div class="text-red-500 mb-4">
+                @foreach (var error in ViewData.ModelState.Values.SelectMany(v => v.Errors))
+                {
+                    <p>@error.ErrorMessage</p>
+                }
+            </div>
+        }
+
         <form asp-action="Login" method="post" class="space-y-4">
-            
+
             <div>
                 <input asp-for="Email" class="w-full px-4 py-2 border rounded-lg focus:outline-none focus:ring-2 focus:ring-gray-400 bg-gray-100 text-gray-900 placeholder-gray-500" placeholder="Email-adres" />
                 <span asp-validation-for="Email" class="text-red-500 text-sm"></span>
             </div>
-            
+
             <div>
                 <input asp-for="Password" type="password" class="w-full px-4 py-2 border rounded-lg focus:outline-none focus:ring-2 focus:ring-gray-400 bg-gray-100 text-gray-900 placeholder-gray-500" placeholder="Wachtwoord" />
                 <span asp-validation-for="Password" class="text-red-500 text-sm"></span>
@@ -36,9 +47,8 @@
                     <input asp-for="RememberMe" class="h-4 w-4 text-gray-800 focus:ring-gray-400 border-gray-300 rounded mr-2" />
                     <label asp-for="RememberMe" class="text-sm text-gray-600">Onthoud mij</label>
                 </div>
-                <a href="/forgot-password" class="text-sm text-gray-500 hover:underline">Wachtwoord vergeten?</a>
             </div>
-            
+
             <div>
                 <button type="submit" class="w-full bg-gray-600 text-white rounded-lg py-2 text-center font-medium hover:bg-gray-700 focus:ring-2 focus:ring-gray-400 focus:ring-opacity-50">
                     Login
