+++ bumbo/Views/Shared/Components/NavBarViewComponent.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using bumbo.Models;
+using bumbo.ViewModels;
 
 namespace bumbo.ViewComponents
 {
@@ -41,12 +42,3 @@ namespace bumbo.ViewComponents
 }
 
 
-public class NavBarViewModel
-{
-        public string FirstName { get; set; }
-        public string Email { get; set; }
-        public bool IsAuthenticated { get; set; }
-        public bool IsSystemManager { get; set; }
-        public bool IsBranchManager { get; set; }
-
-}
